Extract bindings.py from _compute/runtime.py
Replace the re-export shim in braincell/_compute/bindings.py with the real
bodies of the 25 channel-binding symbols, moved verbatim out of runtime.py:
channel/ion binding resolution, the two bound-channel wrappers, runtime node
instantiation, the merged channel-layout install path, per-node parameter
reads and post-compilation write-back, and the subsolver/root-level
classification helpers.

Pure relocation; all 25 bodies are byte-identical to their previous text.
bindings.py imports nothing from runtime.py: it depends downward on ions for
_build_runtime_ions and _sync_runtime_ion and on layouts for MechanismLayout,
outward on braincell.mech, braincell.quad, braincell.ion and braincell.channel,
and refers to CellRuntimeState only through a TYPE_CHECKING import of .state.

Two dead symbols are deleted rather than moved. _quantity_full had no caller
anywhere; its only references were its own def and the shim import list this
commit replaces. _iter_runtime_synapse_layouts had no reference but its own
def; the live near-duplicate method CellRuntimeState.iter_synapse_layouts and
its four callers are untouched.

runtime.py drops the six imports orphaned by the move -- runtime_ion, Channel,
Markov, KineticIon, get_registry and get_integrator -- and re-exports all 25
moved symbols so its attribute surface is unchanged.
